Add unit-test for small sample sizes
