+++ sktree/stats/tests/test_forestht.py
@@ -6,17 +6,11 @@ import pytest
 from flaky import flaky
 from joblib import Parallel, delayed
 from numpy.testing import assert_array_equal
-from scipy.special import expit
 from sklearn import datasets
 
 from sktree import HonestForestClassifier, RandomForestClassifier, RandomForestRegressor
 from sktree._lib.sklearn.tree import DecisionTreeClassifier
-from sktree.stats import (
-    FeatureImportanceForestClassifier,
-    FeatureImportanceForestRegressor,
-    PermutationForestClassifier,
-    PermutationForestRegressor,
-)
+from sktree.stats import FeatureImportanceForestClassifier, FeatureImportanceForestRegressor
 from sktree.stats.utils import _non_nan_samples
 from sktree.tree import ObliqueDecisionTreeClassifier
 
@@ -80,198 +74,23 @@ def test_featureimportance_forest_errors():
         estimator=RandomForestClassifier(
             n_estimators=10,
         ),
+        test_size=0.5,
         permute_per_tree=permute_per_tree,
         sample_dataset_per_tree=sample_dataset_per_tree,
     )
     with pytest.raises(RuntimeError, match="The estimator must be fitted"):
         est.train_test_samples_
 
-    with pytest.raises(RuntimeError, match="There are less than 2 testing samples"):
+    with pytest.raises(RuntimeError, match="There are less than 5 testing samples"):
         est.statistic(iris_X[:5], iris_y[:5])
 
-    est = FeatureImportanceForestClassifier(estimator=RandomForestRegressor)
+    est = FeatureImportanceForestClassifier(estimator=RandomForestRegressor, test_size=0.5)
     with pytest.raises(RuntimeError, match="Estimator must be"):
-        est.statistic(iris_X[:10], iris_y[:10])
+        est.statistic(iris_X[:20], iris_y[:20])
 
-    est = FeatureImportanceForestRegressor(estimator=RandomForestClassifier)
+    est = FeatureImportanceForestRegressor(estimator=RandomForestClassifier, test_size=0.5)
     with pytest.raises(RuntimeError, match="Estimator must be"):
-        est.statistic(iris_X[:10], iris_y[:10])
-
-
-@flaky(max_runs=3)
-@pytest.mark.slowtest
-@pytest.mark.parametrize(
-    "hypotester, model_kwargs, n_samples, n_repeats, test_size",
-    [
-        [
-            PermutationForestRegressor,
-            {
-                "estimator": RandomForestRegressor(
-                    max_features="sqrt",
-                    random_state=seed,
-                    n_estimators=75,
-                    n_jobs=-1,
-                ),
-                "random_state": seed,
-            },
-            300,
-            50,
-            0.1,
-        ],
-        [
-            FeatureImportanceForestRegressor,
-            {
-                "estimator": RandomForestRegressor(
-                    max_features="sqrt",
-                    random_state=seed,
-                    n_estimators=125,
-                    n_jobs=-1,
-                ),
-                "random_state": seed,
-                "permute_per_tree": True,
-                "sample_dataset_per_tree": True,
-            },
-            300,  # n_samples
-            1000,  # n_repeats
-            0.2,  # test_size
-        ],
-    ],
-)
-def test_linear_model(hypotester, model_kwargs, n_samples, n_repeats, test_size):
-    r"""Test hypothesis testing forests using MSE from linear model simulation.
-
-    See https://arxiv.org/pdf/1904.07830.pdf Figure 1.
-
-    Y = Beta * X_1 + Beta * I(X_6 = 2) + \epsilon
-    """
-    beta = 15.0
-    sigma = 0.05
-    metric = "mse"
-
-    # sample covariates
-    X_15 = rng.uniform(0, 1, size=(n_samples, 5))
-    X_610 = np.zeros((n_samples, 5))
-    for idx in range(5):
-        X_610[:, idx] = np.argwhere(
-            rng.multinomial(1, [1.0 / 3, 1.0 / 3, 1.0 / 3], size=(n_samples,))
-        )[:, 1]
-    X = np.concatenate((X_15, X_610), axis=1)
-    assert X.shape == (n_samples, 10)
-
-    # sample noise
-    epsilon = rng.normal(size=n_samples, loc=0.0, scale=sigma)
-
-    # compute final y of (n_samples,)
-    y = beta * X[:, 0] + (beta * (X[:, 5] == 2.0)) + epsilon
-    est = hypotester(test_size=test_size, **model_kwargs)
-
-    # test for X_1
-    stat, pvalue = est.test(X, y, [0], metric=metric, n_repeats=n_repeats)
-    print("X1: ", pvalue)
-    assert pvalue < 0.05, f"pvalue: {pvalue}"
-
-    # test for X_6
-    stat, pvalue = est.test(X, y, [5], metric=metric, n_repeats=n_repeats)
-    print("X6: ", pvalue)
-    assert pvalue < 0.05, f"pvalue: {pvalue}"
-
-    # test for a few unimportant other X
-    for covariate_index in [1, 6]:
-        # test for X_2, X_7
-        stat, pvalue = est.test(X, y, [covariate_index], metric=metric, n_repeats=n_repeats)
-        print("X2/7: ", pvalue)
-        assert pvalue > 0.05, f"pvalue: {pvalue}"
-
-
-@flaky(max_runs=3)
-@pytest.mark.slowtest
-@pytest.mark.parametrize(
-    "hypotester, model_kwargs, n_samples, n_repeats, test_size",
-    [
-        [
-            PermutationForestClassifier,
-            {
-                "estimator": RandomForestClassifier(
-                    max_features="sqrt",
-                    random_state=seed,
-                    n_estimators=50,
-                    n_jobs=-1,
-                ),
-                "random_state": seed,
-            },
-            600,
-            50,
-            1.0 / 6,
-        ],
-        [
-            # XXX: Currently does not work with permute and sample dataset per tree
-            FeatureImportanceForestClassifier,
-            {
-                "estimator": RandomForestClassifier(
-                    max_features="sqrt",
-                    random_state=seed,
-                    n_estimators=100,
-                    n_jobs=-1,
-                ),
-                "random_state": seed,
-                "permute_per_tree": False,
-                "sample_dataset_per_tree": False,
-            },
-            600,  # n_samples
-            1000,  # n_repeats
-            1.0 / 6,  # test_size
-        ],
-    ],
-)
-def test_correlated_logit_model(hypotester, model_kwargs, n_samples, n_repeats, test_size):
-    r"""Test MIGHT using MSE from linear model simulation.
-
-    See https://arxiv.org/pdf/1904.07830.pdf Figure 1.
-
-    P(Y = 1 | X) = expit(beta * \\sum_{j=2}^5 X_j)
-    """
-    beta = 10.0
-    metric = "mse"
-
-    n = 100  # Number of time steps
-    ar_coefficient = 0.015
-
-    X = np.zeros((n_samples, n))
-    for idx in range(n_samples):
-        # sample covariates
-        white_noise = rng.standard_normal(size=n)
-
-        # Create an array to store the simulated AR(1) time series
-        ar1_series = np.zeros(n)
-        ar1_series[0] = white_noise[0]
-
-        # Simulate the AR(1) process
-        for t in range(1, n):
-            ar1_series[t] = ar_coefficient * ar1_series[t - 1] + white_noise[t]
-
-        X[idx, :] = ar1_series
-
-    # now compute the output labels
-    y_proba = expit(beta * X[:, 1:5].sum(axis=1))
-    assert y_proba.shape == (n_samples,)
-    y = rng.binomial(1, y_proba, size=n_samples)  # .reshape(-1, 1)
-
-    est = hypotester(test_size=test_size, **model_kwargs)
-
-    # test for X_2 important
-    stat, pvalue = est.test(X.copy(), y.copy(), [1], n_repeats=n_repeats, metric=metric)
-    print("X2: ", pvalue)
-    assert pvalue < 0.05, f"pvalue: {pvalue}"
-
-    # test for X_1 unimportant
-    stat, pvalue = est.test(X.copy(), y.copy(), [0], n_repeats=n_repeats, metric=metric)
-    print("X1: ", pvalue)
-    assert pvalue > 0.05, f"pvalue: {pvalue}"
-
-    # test for X_500 unimportant
-    stat, pvalue = est.test(X.copy(), y.copy(), [n - 1], n_repeats=n_repeats, metric=metric)
-    print("X500: ", pvalue)
-    assert pvalue > 0.05, f"pvalue: {pvalue}"
+        est.statistic(iris_X[:20], iris_y[:20])
 
 
 @flaky(max_runs=2)
@@ -398,10 +217,10 @@ def test_forestht_check_inputs(forest_hyppo):
 
 
 @pytest.mark.parametrize("backend", ["loky", "threading"])
-@pytest.mark.parametrize("n_jobs", [1, -1])
+@pytest.mark.parametrize("n_jobs", [1, 2])
 def test_parallelization(backend, n_jobs):
     """Test parallelization of training forests."""
-    n_samples = 100
+    n_samples = 20
     n_features = 5
     X = rng.uniform(size=(n_samples, n_features))
     y = rng.integers(0, 2, size=n_samples)  # Binary classification
@@ -529,3 +348,27 @@ def test_permute_per_tree_samples_consistency_with_sklearnforest():
 
     assert_array_equal(clf.train_test_samples_[0][0], other_clf.train_test_samples_[0][0])
     assert_array_equal(clf.train_test_samples_[0][1], other_clf.train_test_samples_[0][1])
+
+
+def test_small_dataset():
+    n_samples = 32
+    n_features = 5
+    X = rng.uniform(size=(n_samples, n_features))
+    y = rng.integers(0, 2, size=n_samples)  # Binary classification
+
+    clf = FeatureImportanceForestClassifier(
+        estimator=HonestForestClassifier(
+            n_estimators=10, random_state=seed, n_jobs=1, honest_fraction=0.5
+        ),
+        test_size=0.2,
+        permute_per_tree=False,
+        sample_dataset_per_tree=False,
+    )
+    stat, pvalue = clf.test(X, y, covariate_index=[1, 2], metric="mi")
+    assert ~np.isnan(pvalue)
+    assert ~np.isnan(stat)
+    assert pvalue > 0.05
+
+    stat, pvalue = clf.test(X, y, metric="mi")
+    assert stat == 0.0
+    assert pvalue > 0.05
